backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0
django-filter==24.3
psycopg[binary]==3.2.1
python-dotenv==1.0.1
opensearch-py==2.6.0
pandas==2.2.2
duckdb==1.1.1
prefect==3.0.0
mlflow==2.16.0
scikit-learn==1.5.2
prometheus-client==0.20.0
gunicorn==22.0.0
